Create Project: Connect Four
Connect Four is a fun game for two players. The players place discs on a vertical board that is 7 columns wide and 6 rows tall. The goal of the game is to be the first to form a horizontal, vertical, or diagonal line of four discs. In this project, you will develop this game with a number of improvements.